Changes in the ipfs-search UI
Simplify the webchannel communication for the ipfs-search
UI. Before, all results would appear and later on the ones
with no providers would be disabled, now this is done
on-the-fly.

Fix contractDeploy() and new blocks watch task

Implement the 'view page source' action in the browser.

Replace the remaining wait_for() calls in ipfsops