Use a single script for injection scripts
This should fix the case where window.ipfs was set
only just after the DOM creation. Because it seems
we can't control the order in which QtWebEngine executes
the scripts in the scripts collection, this is the
working solution for now.

* ipfs.encrypt: coroutine to decrypt a JSON stored in the MFS
* ui.browser: show a status page when switching profiles
* setup.py: make it possible to disable 'connectSlotsByName()'
(from QMetaObject) for certain forms when running pyuic
* webprofiles: make the default web profile configurable in the
settings

revbump to 0.4.11