Implement automatic CIDv0 -> CIDv1 upgrade in IPFSPath
* ipfs.cidhelpers.IPFSPath: add an autoCidConv kw argument
that enables automatic CID upgrade (not enabled by default).
It's used in core places (like the resource opener) to have
transparent CID upgrades in most of the application

* implement automatic CIDv0 upgrade for the native scheme
when URL input in the browser has a ipfs://<cidv0-base58>/..
format (which would normally be invalid). The upgrade had to
be done early in handleEditedUrl() before the QUrl() is
instantiated (otherwise we lose the case). Suggested by @lidel
in issue 5.

* implement pinning of the page's links, and pinning of the
parent node

* early implementation of the quick-access q:// scheme

* ipfs.pinning: change the return value of the pin coroutine

* ipfs.distipfsfetch: fetch go-ipfs 0.4.22 by default, and use
that version as well in the AppImage and DMG

revbump to 0.4.12